requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# AI / LLM
anthropic>=0.21.0
openai>=1.12.0

# Memory / Database
supabase>=2.4.0
httpx>=0.27.0

# Voice (optional)
# openwakeword>=0.6.0  # Wake word detection
# faster-whisper>=1.0.0  # Local STT
# piper-tts>=1.2.0  # Local TTS

# Vision
Pillow>=10.2.0
mss>=9.0.1  # Screenshots

# Device Control
pyautogui>=0.9.54
playwright>=1.42.0

# Web Search
duckduckgo-search>=5.0.0

# Utilities
aiofiles>=23.2.1
python-multipart>=0.0.9
schedule>=1.2.1
rich>=13.7.0
typer>=0.9.0
